refactor(direct-control): extract session config
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves resolveCiv7DirectControlConfig and its private host/env/port helpers into src/session/config.ts while keeping the public facade export in index.ts.

Preserves host/env ordering, deduplication, port and timeout defaults, invalid-port classification, and existing CLI resolver behavior. This does not change endpoint discovery, socket/session lifecycle, command execution, reconnect behavior, health polling, or runtime-status projection. Endpoint discovery, session socket/reconnect source ownership, public procedure schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- session.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check
- ownership scan for session config helper definitions

Local package/source relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
